+++ app/src/main/java/com/babytracker/designsystem/components/button/ButtonDefaults.kt
@@ -37,6 +37,8 @@ object ButtonDefaults {
     @Composable fun contentColor(): Color = LocalAppComponentTokens.current.button.contentColor
     @Composable fun disabledContainerColor(): Color = LocalAppComponentTokens.current.button.disabledContainerColor
     @Composable fun disabledContentColor(): Color = LocalAppComponentTokens.current.button.disabledContentColor
+    @Composable fun tonalContainerColor(): Color = LocalAppComponentTokens.current.button.tonalContainerColor
+    @Composable fun tonalContentColor(): Color = LocalAppComponentTokens.current.button.tonalContentColor
     @Composable fun secondaryContentColor(): Color = LocalAppComponentTokens.current.button.secondaryContentColor
     @Composable fun textContentColor(): Color = LocalAppComponentTokens.current.button.textContentColor
     @Composable fun dangerContainerColor(): Color = LocalAppComponentTokens.current.button.dangerContainerColor